Move language name getters into a trait
so that we can use them inside any Entity having
a language code attribute, allowing to pass
language names to the views within json data.
